Revamp token handling and tiling
We don’t store tokens and comment attributes in the `NSMutableAttributedString` of `NSTextStorage` anymore. Instead, we maintain that information as part of the line map and synthesise attribute information (e.g., for syntax highlighting) on the fly.

We are also more careful with tiling (on macOS) to avoid invalidating information and triggering recomputes or redraws.

Contributes to #24, #29, #13, and #57