Different minimap layout
We used to have a custom text content manager for the minimap that changes the font size of text in text elements. This required an ultimately unsafe construction with two text content managers operating on one text storage.

The new setup uses two text layout manager on a single text content manager with the layour manager of the code view being the primary one. This is an officially supported setup, and hence, safer.

As a result, the minimap performs layout with the same sizes as the code view, but scales everything down after layout, before rendering.